fix crush on android 4.x

+++ app/src/main/java/ru/solandme/washwait/POJO/weather/Sys.java
@@ -14,10 +14,10 @@ public class Sys {
     private String country;
     @SerializedName("sunrise")
     @Expose
-    private int sunrise;
+    private long sunrise;
     @SerializedName("sunset")
     @Expose
-    private int sunset;
+    private long sunset;
 
     public double getMessage() {
         return message;
@@ -35,19 +35,19 @@ public class Sys {
         this.country = country;
     }
 
-    public int getSunrise() {
+    public long getSunrise() {
         return sunrise;
     }
 
-    public void setSunrise(int sunrise) {
+    public void setSunrise(long sunrise) {
         this.sunrise = sunrise;
     }
 
-    public int getSunset() {
+    public long getSunset() {
         return sunset;
     }
 
-    public void setSunset(int sunset) {
+    public void setSunset(long sunset) {
         this.sunset = sunset;
     }
